Add a "dummy" consumer to skip through old data.
It would be better to directly manipulate the consumer offsets to
skip over data, but failing that, the DummyConsumer can be used to
process a topic+group with a "no-op" operation.

Note that this consumes a lot of bandwidth on the Bagheera nodes,
so it may impact other consumers too.